+++ includes/platformer-engine/Engine/Engine.hpp
@@ -3,7 +3,8 @@
 
 #include "Render/Window.hpp"
 #include "Timer/Timer.hpp"
-#include "../../../src/Physics/PhysicsSystem.hpp"
+#include "Physics/PhysicsSystem.hpp"
+#include "GameLevel/GameLevel.hpp"
 #include <memory>