+++ src/Render/Window.cpp
@@ -1,19 +1,23 @@
 #include "Render/Window.hpp"
+#include "../../includes/spic-prj-api/Input.hpp"
+#include "LevelParser/LevelParser.hpp"
 
 platformer_engine::Window::Window(int width, int height, const std::string &title, const spic::Color &color) {
-    _graphicsFacade.Init(width, height, title, color);
+    _graphicsFacade = std::make_shared<GraphicsFacade>();
+    _graphicsFacade->Init(width, height, title, color);
+    TextureManager::GetInstance().Init(_graphicsFacade);
 }
 
 void platformer_engine::Window::Render() {
-    _graphicsFacade.Clear();
-    _graphicsFacade.Render();
+    _graphicsFacade->Clear();
+    _graphicsFacade->Render();
 }
 
 void platformer_engine::Window::Quit() {
-    _graphicsFacade.Quit();
+    _graphicsFacade->Quit();
 }
 
 auto platformer_engine::Window::ListenForEvents() -> std::vector<EventsEnum> {
-   return _inputFacade.ListenForInput();
+   return InputFacade::GetInstance().ListenForInput();
 }